CacheDriver: moved to subdirectory
